AutoLoad: Fix support for feature-based overrides
Current status quo was that overridden autoloads are not supported
since the AutoLoad creation panel doesn't support adding dots in
filenames (as they will be used as node names).

Still, 24453 (intentionally not making a hyperlink to avoid spam
while this is WIP) shows that overrides can be added from the General
tab, and work, with caveats. This commit solves the caveats.

Fixes 24453.