WIP: Replace etc2comp and squish by etcpak for compression
etcpak is a *much* faster alternative to our previous libraries for both
ETC/ETC2 and DXT1/DXT5 formats.

Backport `Image::get_image_mipmap_offset_and_dimensions` from `master`.

---

WIP as it turns out we can't rely solely on etcpak for GLES2 support:

<reduz> the problem is that you dont have channel remapping in GLES2, so you
need the support for BC4 and BC5. In GLES3 and Vulkan this is not a problem
because you can use BC1-3 with channel remapping.
<reduz> so, since ETCPack does not support BC4 an 5, its kind of useless there

Could make some hybrid solution where we use etcpak for the supported formats
and keep squish and etc2comp for the rest. Requires more work, might not be
worth it when it's already solved for 4.0.