SCons: Add optional ninja build tool (`use_ninja=yes`)
The tool generates a `ninja.build` file which can be used by the `ninja` build
system to compile the codebase.

This tool is being worked on upstream and should eventually make it to a stable
SCons release, and will then supersede this project-localized one.

Requires SCons 4.2+, the `ninja` build system and `ninja` Python module.

Usage: `scons use_ninja=yes` then `ninja` to build.