CI: Speed up static checks by checking only changed files
file_format, header_guards and clang-format benefit from this short list.
dotnet-format and the JS checks don't, but they're only relevant for PRs
changing a specific set of files, so we skip them when those files aren't
modified.

The logic to get changed files only works reliably for:
- Pull request events
- Non-force pushed push events

So when force pushing a branch in your fork, or creating a new branch,
it will still scan all files as fallback.

Upgraded CI runner to Ubuntu 22.04 so we get clang-format 14 out of the box,
so we don't need to install a custom version (saves ~15 s).