Fix typed array export
Fixes to changes from #76114.

This PR remove the hack with pointer and replace it with a proper handling
from EditorPropertyNodePath of both NodePath and Node. Also needed to revise
a bit the saving and loading of scene to handle properly the change.

Fixes #74050.
Fixes #74141.
Fixes #75981.
Fixes #76642.

Revert "Fix typed array export... again"

This reverts commit da8d6734fbc31f68e7e822f37fd239a92ac79b34.

Co-Authored-By: Tomek <kobewi4e@gmail.com>