Remove denoise module and thirdparty OIDN
This is replaced by a much lighter weight and faster JNLM denoiser.

OIDN is still much more accurate, and may be provided as an optional backend
in the future, but the JNLM denoiser seems good enough for most use cases and
remove OIDN reduces the buildsystem complexity, binary size, and build times
for Godot very significantly.