Add support for SDL3 joystick input driver
Made possible by EIREXE, xsellier and the SDL team.

This commit includes statically linked SDL3 for Windows, Linux and macOS.
The vendored copy of SDL3 was setup to only build the requirement subsystems,
with some patches to be able to make it as minimal as possible and reduce
the impact on binary size and code size.

Co-Authored-By: Álex Román Núñez <eirexe123@gmail.com>
Co-Authored-By: Xavier Sellier <xsellier@gmail.com>